Conflicts:
	components/insight/SRC/org/openmicroscopy/shoola/env/ui/SplashScreenManager.java

+++ components/insight/SRC/org/openmicroscopy/shoola/agents/measurement/view/MeasurementViewerUI.java
@@ -69,6 +69,7 @@ import org.openmicroscopy.shoola.agents.events.measurement.SelectPlane;
 import org.openmicroscopy.shoola.agents.measurement.IconManager;
 import org.openmicroscopy.shoola.agents.measurement.MeasurementAgent;
 import org.openmicroscopy.shoola.agents.measurement.actions.MeasurementViewerAction;
+import org.openmicroscopy.shoola.agents.measurement.actions.UnitsAction;
 import org.openmicroscopy.shoola.agents.util.EditorUtil;
 import pojos.WorkflowData;
 import org.openmicroscopy.shoola.env.config.Registry;
@@ -469,6 +470,10 @@ class MeasurementViewerUI
         		IconManager.MEASUREMENT_TOOL);
         if (icon != null) setIconImage(icon.getImage());
         initComponents();
+        UnitsAction a = (UnitsAction)
+        	controller.getAction(MeasurementViewerControl.IN_MICRONS);
+    	a.setRefUnits(EditorUtil.transformSize(
+    			model.getPixelSizeX()).getUnits());
         //buildGUI();
     }
     
@@ -523,25 +528,33 @@ class MeasurementViewerUI
 			model.notifyDataChanged(true);
 			ROI newROI = model.cloneROI(idList.get(0));
 			ROIShape newShape;
+			ROIFigure roi;
 			for (ROIShape shape : shapeList)
 			{
-				newShape = new ROIShape(newROI, shape.getCoord3D(), shape);
-				if (getDrawing().contains(shape.getFigure()))
-				{
-					shape.getFigure().removeFigureListener(controller);
-					getDrawing().removeDrawingListener(controller);
-					getDrawing().remove(shape.getFigure());
-					getDrawing().addDrawingListener(controller);
+				roi = shape.getFigure();
+				if (!roi.isReadOnly() && roi.canEdit()) {
+					newShape = new ROIShape(newROI, shape.getCoord3D(), shape);
+					if (getDrawing().contains(roi))
+					{
+						if (roi.canAnnotate())
+							roi.removeFigureListener(controller);
+						getDrawing().removeDrawingListener(controller);
+						getDrawing().remove(shape.getFigure());
+						getDrawing().addDrawingListener(controller);
+					}
+					model.deleteShape(shape.getID(), shape.getCoord3D());
+					if (newShape.getCoord3D().equals(model.getCurrentView()))
+					{
+						getDrawing().removeDrawingListener(controller);
+						getDrawing().add(newShape.getFigure());
+						if (roi.canAnnotate())
+							newShape.getFigure().addFigureListener(
+										controller);
+						getDrawing().addDrawingListener(controller);
+					}
+					model.addShape(newROI.getID(), newShape.getCoord3D(),
+							newShape);
 				}
-				model.deleteShape(shape.getID(), shape.getCoord3D());
-				if (newShape.getCoord3D().equals(model.getCurrentView()))
-				{
-					getDrawing().removeDrawingListener(controller);
-					getDrawing().add(newShape.getFigure());
-					newShape.getFigure().addFigureListener(controller);
-					getDrawing().addDrawingListener(controller);
-				}
-				model.addShape(newROI.getID(), newShape.getCoord3D(), newShape);
 			}
 		}
 		catch (Exception e)
@@ -569,25 +582,32 @@ class MeasurementViewerUI
 			model.notifyDataChanged(true);
 			ROI newROI = model.cloneROI(id);
 			ROIShape newShape;
+			ROIFigure roi;
 			for (ROIShape shape : shapeList)
 			{
-				newShape = new ROIShape(newROI, shape.getCoord3D(), shape);
-				if (getDrawing().contains(shape.getFigure()))
-				{
-					shape.getFigure().removeFigureListener(controller);
-					getDrawing().removeDrawingListener(controller);
-					getDrawing().remove(shape.getFigure());
-					getDrawing().addDrawingListener(controller);
+				roi = shape.getFigure();
+				if (!roi.isReadOnly() && roi.canEdit()) {
+					newShape = new ROIShape(newROI, shape.getCoord3D(), shape);
+					if (getDrawing().contains(roi))
+					{
+						if (roi.canAnnotate())
+							roi.removeFigureListener(controller);
+						getDrawing().removeDrawingListener(controller);
+						getDrawing().remove(shape.getFigure());
+						getDrawing().addDrawingListener(controller);
+					}
+					model.deleteShape(shape.getID(), shape.getCoord3D());
+					if (newShape.getCoord3D().equals(model.getCurrentView()))
+					{
+						getDrawing().removeDrawingListener(controller);
+						this.getDrawing().add(newShape.getFigure());
+						if (roi.canAnnotate())
+							newShape.getFigure().addFigureListener(controller);
+						getDrawing().addDrawingListener(controller);
+					}
+					model.addShape(newROI.getID(), newShape.getCoord3D(),
+							newShape);
 				}
-				model.deleteShape(shape.getID(), shape.getCoord3D());
-				if (newShape.getCoord3D().equals(model.getCurrentView()))
-				{
-					getDrawing().removeDrawingListener(controller);
-					this.getDrawing().add(newShape.getFigure());
-					newShape.getFigure().addFigureListener(controller);
-					getDrawing().addDrawingListener(controller);
-				}
-				model.addShape(newROI.getID(), newShape.getCoord3D(), newShape);
 			}
 		}
 		catch (Exception e)
@@ -616,17 +636,22 @@ class MeasurementViewerUI
 			ROI newROI = model.cloneROI(id);
 			ROIShape newShape;
 			Drawing drawing = model.getDrawing();
+			ROIFigure roi;
 			for (ROIShape shape : shapeList)
 			{
-				newShape = new ROIShape(newROI, shape.getCoord3D(), shape);
-				if (newShape.getCoord3D().equals(model.getCurrentView()))
-				{
-					drawing.removeDrawingListener(controller);
-					drawing.add(newShape.getFigure());
-					newShape.getFigure().addFigureListener(controller);
-					drawing.addDrawingListener(controller);
+				roi = shape.getFigure();
+				if (!roi.isReadOnly()) {
+					newShape = new ROIShape(newROI, shape.getCoord3D(), shape);
+					if (newShape.getCoord3D().equals(model.getCurrentView()))
+					{
+						drawing.removeDrawingListener(controller);
+						drawing.add(newShape.getFigure());
+						newShape.getFigure().addFigureListener(controller);
+						drawing.addDrawingListener(controller);
+					}
+					model.addShape(newROI.getID(), newShape.getCoord3D(),
+							newShape);
 				}
-				model.addShape(newROI.getID(), newShape.getCoord3D(), newShape);
 			}
 			updateDrawingArea();
 		}
@@ -634,6 +659,20 @@ class MeasurementViewerUI
 		{
 			handleROIException(e, CREATE_MSG);
 		}
+	}
+	
+	void markROIForDelete(ROIFigure roi)
+	{
+		if (roi == null) return;
+		long id = roi.getROIShape().getID();
+		if (id < 0) return;
+		try {
+			//model.deleteShape(id, roi.getROIShape().getCoord3D());
+		} catch (Exception e) {
+			
+		}
+		
+		model.markROIForDelete(id, roi.getROI(), true);
 	}
 	
 	/**
@@ -651,17 +690,19 @@ class MeasurementViewerUI
 			for (ROIShape shape : shapeList)
 			{
 				roi = shape.getFigure();
-				r = roi.getROI();
-				if (!r.isClientSide()) b = true;
-				if (getDrawing().contains(roi))
-				{
-					shape.getFigure().removeFigureListener(controller);
-					getDrawing().removeDrawingListener(controller);
-					getDrawing().remove(roi);
-					getDrawing().addDrawingListener(controller);
+				if (roi.canDelete()) {
+					r = roi.getROI();
+					if (!r.isClientSide()) b = true;
+					if (getDrawing().contains(roi))
+					{
+						shape.getFigure().removeFigureListener(controller);
+						getDrawing().removeDrawingListener(controller);
+						getDrawing().remove(roi);
+						getDrawing().addDrawingListener(controller);
+					}
+					model.deleteShape(shape.getID(), shape.getCoord3D());
+					model.markROIForDelete(shape.getID(), r, false);
 				}
-				model.deleteShape(shape.getID(), shape.getCoord3D());
-				model.markROIForDelete(shape.getID(), r);
 			}
 			model.notifyDataChanged(b);
 		} catch (Exception e) {
@@ -1199,13 +1240,15 @@ class MeasurementViewerUI
 							 while (f.hasNext()) {
 								 figure = f.next();
 								 drawing.add(figure);
-								 figure.addFigureListener(controller);
+								 if (figure.canAnnotate())
+									 figure.addFigureListener(controller);
 							 }
 							 f = second.iterator();
 							 while (f.hasNext()) {
 								 figure = f.next();
 								 drawing.add(figure);
-								 figure.addFigureListener(controller);
+								 if (figure.canAnnotate())
+								 	figure.addFigureListener(controller);
 							 }
 						 }
 					} else {
@@ -1224,7 +1267,8 @@ class MeasurementViewerUI
 								{
 									figure = shape.getFigure();
 									drawing.add(figure);
-									figure.addFigureListener(controller);
+									if (figure.canAnnotate())
+										figure.addFigureListener(controller);
 								}
 							}
 						}
@@ -1265,13 +1309,15 @@ class MeasurementViewerUI
 				 while (f.hasNext()) {
 					 figure = f.next();
 					 drawing.add(figure);
-					 figure.addFigureListener(controller);
+					 if (figure.canAnnotate())
+						 figure.addFigureListener(controller);
 				 }
 				 f = second.iterator();
 				 while (f.hasNext()) {
 					 figure = f.next();
 					 drawing.add(figure);
-					 figure.addFigureListener(controller);
+					 if (figure.canAnnotate())
+						 figure.addFigureListener(controller);
 				 }
 			}
 		}
@@ -1289,30 +1335,37 @@ class MeasurementViewerUI
 	 */
 	void propagateShape(ROIShape shape, int timePoint, int zSection) 
 	{
-		List<ROIShape> addedShapes;
+		List<ROIShape> addedShapes = new ArrayList<ROIShape>();
 		try
 		{
-			addedShapes = model.propagateShape(shape, timePoint, zSection);
+			ROIFigure roi = shape.getFigure();
+			if (!roi.isReadOnly() && roi.canEdit())
+				addedShapes = model.propagateShape(shape, timePoint, zSection);
 			ROIFigure figToDelete = null;
 			ROIFigure roiFig;
 			for (ROIShape newShape : addedShapes)
 			{
-				if (newShape.getCoord3D().equals(model.getCurrentView()))
-				{
-					getDrawing().removeDrawingListener(controller);
-					figToDelete = null;
-					for (Figure f : getDrawing().getFigures()) {
-						roiFig = (ROIFigure) f;
-						if (roiFig.getROI().getID() == newShape.getID())
-							figToDelete = roiFig;
+				roi = newShape.getFigure();
+				if (!roi.isReadOnly() && roi.canEdit()) {
+					if (newShape.getCoord3D().equals(model.getCurrentView()))
+					{
+						getDrawing().removeDrawingListener(controller);
+						figToDelete = null;
+						for (Figure f : getDrawing().getFigures()) {
+							roiFig = (ROIFigure) f;
+							if (roiFig.getROI().getID() == newShape.getID())
+								figToDelete = roiFig;
+						}
+						if (figToDelete!=null)
+							getDrawing().remove(figToDelete);
+						this.getDrawing().add(newShape.getFigure());
+						if (roi.canAnnotate())
+							newShape.getFigure().addFigureListener(controller);
+						getDrawing().addDrawingListener(controller);
 					}
-					if (figToDelete!=null)
-						getDrawing().remove(figToDelete);
-					this.getDrawing().add(newShape.getFigure());
-					newShape.getFigure().addFigureListener(controller);
-					getDrawing().addDrawingListener(controller);
+					newShape.getFigure().calculateMeasurements();
 				}
-				newShape.getFigure().calculateMeasurements();
+				
 			}
 			if (!model.isHCSData()) roiManager.addROIShapes(addedShapes);
 		}
@@ -1413,7 +1466,7 @@ class MeasurementViewerUI
      * @return See above.
      */
     long getPixelsID() { return model.getPixelsID(); }
- 
+    
     /**
 	 * Calculate the stats for the Rois in the shapelist. This method
 	 * will call the graphView.
@@ -1561,10 +1614,10 @@ class MeasurementViewerUI
 		ROIFigure figure;
 		while (i.hasNext()) {
 			figure = i.next();
-			if (!figure.isReadOnly()) {
+			//if (!figure.isReadOnly()) {
 				MeasurementAttributes.SHOWTEXT.set(figure, show);
 				if (roiInspector != null) roiInspector.showText(show, figure);
-			}
+			//}
 		}
 		model.getDrawingView().repaint();
 	}
@@ -1601,6 +1654,8 @@ class MeasurementViewerUI
     public void setOnScreen()
     {
     	setSize(DEFAULT_SIZE);
+    	UIUtilities.incrementRelativeToAndShow(null, this);
+    	/*
         if (model != null) { //Shouldn't happen
             UIUtilities.setLocationRelativeToAndSizeToWindow(
             		model.getRequesterBounds(), this, MAXIMUM_SIZE);
@@ -1608,6 +1663,7 @@ class MeasurementViewerUI
             //pack();
             UIUtilities.incrementRelativeToAndShow(null, this);
         }
+        */
     }
     
 	/**
